Move LRUCache into impl
